Optimize

+++ User/Function.h
@@ -13,7 +13,7 @@ void ShiftLED(void);
 void SendData(void);
 void ClearData(void);
 void Delay(uint32_t TimeDelay);
-uint8_t StartPos(uint16_t Pos);
+uint8_t StartPos(uint32_t Pos);
 void DisplayOneIC(uint8_t Data);
 void DisplayLine(uint8_t U1, uint8_t U2, uint8_t U3, uint8_t U4, uint8_t U5);
 uint8_t CharCVT(char c, uint8_t times);;
